Fix: MailQueue.schedule_retry_in() based next retry on previous attempt
Another retry was scheduled with specified offset from the _scheduled_
time of previous retry, not the actual one.

In order to test these kinds of scenarios I had to pull
timestamp-related logic from the database (e.g. I no longer rely on
DB's default value for next_attempt_at column). It's more brittle in
theory, but the DB is currently embedded in the same process (thus
sharing the clock), and this kind of setup allows to extensively test
various timing scenarios, soI think it's worth it in the current
context.